+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Features Added
 
 - Added enum `AggregationType` which can be used to specify aggregations in the query API.
+- Added `LogsBatchQueryResult` model that is returned for a logs batch query.
 
 ### Breaking Changes
 
@@ -13,6 +14,7 @@
 - `batch_query` API now returns a list of responses.
 - `LogsBatchResults` model is now removed.
 - `LogsQueryRequest` is renamed to `LogsBatchQueryRequest`
+- `LogsQueryResults` is now renamed to `LogsQueryResult`
 
 ### Bugs Fixed